+++ src/components/Header.scss
@@ -1,0 +1,29 @@
+$color-blue: #127ccb;
+$color-gray: #dbdbdb;
+
+.header {
+  background-color: $color-blue;
+  padding: 30px 40px;
+
+  .search-status {
+    max-width: 980px;
+    margin: 0px auto;
+
+    color: white;
+    font-size: 24px;
+    font-weight: 700;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+
+@media (max-width: 600px) {
+  .header {
+    .search-status {
+      display: initial;
+      text-align: center;
+    }
+  }
+}
