+++ src/components/index.js
@@ -2,5 +2,6 @@
  * @author Shiming Chen <chen@lemontv.me>
  */
 
-export { default as App } from './App';
+export { default, default as App } from './App';
 export { default as Departures } from './Departures';
+export { default as Header } from './Header';
