+++ src/containers/index.js
@@ -1,2 +1,3 @@
-export { default as App } from './App';
+export { default, App } from 'components'
 export { default as Departures } from './Departures';
+export { default as Header } from './Header';
